games/stuntrally: Switch to binary release until ported to ogre-1.11+
StuntRally does not currently work on ogre-1.11+. Attempts at adding
patches allows compiling, but leads to a segfault when launching.
This is a temporary measure to allow ogre to be upgraded to a newer
version without breaking this program. Once the port to 1.11+ is
completed, StuntRally will be reverted back to a source compile.