network/teamviewer: Readded and updated for version 15.25.5
It looks like teamviewer requires it license files to be located in
/opt/teamviewer/doc so moving these files to the standard slackware path
(/usr/doc/$PROGNAM-$VERSION) prevents it from working.

Here I basically do not move these files but we could also create a
symlink between the slack path and the teamviewer path. Not sure which
would be prefered.

Also I am not particularly asking to take over the maintenance of this.
just that I managed to fix it on all my machines and thought I would
submit it back for testing and discussion.

Signed-off-by: ArTourter <artourter@gmail.com>