multimedia/kodi: Add README on python issues.
There have been issues with kodi addons causing kodi to crash. It has been
determined that the problem is with python3 including the version that was
released with 15.0 and continues through 3.9.12 (the latest patch). This has
been fixed in 3.9.13, but that updated version hasn't made it to 15.0 yet. The
README.python contains instructions on how users can upgrade their python3
package until Pat releases a patch. The instructions will also be saved in the
program's /usr/doc directory.