Disable building tests.
This commit disables building tests for mutter and gnome-shell, so that
mutter can drop the dependency on xvfb-run. The gnome-shell slackbuild
is included in this commit, since it must disable building tests as well,
otherwise gnome-shell will fail-to-build when it tries to build its
mutter tests.

I've tested building and runtime operation of the reverse dependencies
of this change, which are gnome-shell, and the budgie-desktop packages.
Everything is building and working as expected.